requirements.txt
python-binance==1.0.19
pandas>=1.3.0
numpy>=1.20.0
